Implements Criteria class
