+++ src/main/java/com/jalasoft/webservice/controller/Criteria.java
@@ -1,0 +1,13 @@
+package com.jalasoft.webservice.controller;
+
+public class Criteria {
+    private String filePath;
+
+    public String getFilePath() {
+        return filePath;
+    }
+
+    public void setFilePath(String filePath) {
+        this.filePath = filePath;
+    }
+}
